requirements.txt
django>=4.2,<5.0
django-jazzmin>=3.0,<4.0
django-ckeditor==6.7.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
whitenoise>=6.6,<7.0
Pillow>=10.0,<11.0
